backend/requirements.txt
# --- Framework Principal ---
Flask==3.0.3
Flask-Cors==4.0.1
Werkzeug==3.0.3

# --- Banco de Dados ---
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg==3.1.19 # MODERNO: Substitui o psycopg2-binary, é mais rápido e fácil de instalar.

# --- Utilitários ---
python-dotenv==1.0.1
gunicorn==22.0.0 # Essencial para produção em ambientes Linux/macOS.

# --- API e Autenticação ---
Flasgger==0.9.7.1
PyJWT==2.8.0

# --- Comunicação em Tempo Real (para a próxima etapa) ---
paho-mqtt==2.1.0

# --- SSL/TLS ---
certifi==2024.8.30
